fastapi_backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=13.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.9
livekit-agents[openai,silero,elevenlabs,groq]>=1.0.11
livekit-agents[google]>=1.0.0
livekit-plugins-noise-cancellation>=0.2.0
livekit-api>=1.0.0
serpapi>=0.1.5
aiohttp>=3.9.0
google-generativeai>=0.8.0
requests>=2.32.0
spotipy>=2.25.0
lyricsgenius>=3.6.0